Refactor task exports into index.
